Applied migrations to DB and express session management along with few other issues

+++ controllers/posts.controller.js
@@ -9,11 +9,17 @@ exports.createPost = (req, res) => {
     });
     return;
   }
+  if ((req.session==undefined || req.session.user==undefined)) {
+    res.status(400).send({
+      message: "Please login first!"
+    });
+    return;
+  }
   const post = {
     title: req.body.title,
     content: req.body.content,
     published: req.body.published,
-    userId: req.body.userId,
+    userId: req.session.user.id,
   }
   posts.create(post)
     .then((post) => {
@@ -28,7 +34,13 @@ exports.createPost = (req, res) => {
 
 
 exports.findAllPostsByUserId = (req, res) => {
-  const id = req.body.userId;
+  if ((req.session==undefined || req.session.user==undefined)) {
+    res.status(400).send({
+      message: "Please login first!"
+    });
+    return;
+  }
+  const id = req.session.user.id;
   posts.findAll({ where: { userId: id } })
     .then((post) => {
       if (post) {
@@ -48,7 +60,13 @@ exports.findAllPostsByUserId = (req, res) => {
 };
 
 exports.findPublishedPostsByUserId = (req, res) => {
-  const id = req.body.userId;
+  if ((req.session==undefined || req.session.user==undefined)) {
+    res.status(400).send({
+      message: "Please login first"
+    });
+    return;
+  }
+  const id = req.session.user.id;
   posts.findAll({ where: { userId: id, published: true } })
     .then((post) => {
       if (post) {
@@ -68,7 +86,13 @@ exports.findPublishedPostsByUserId = (req, res) => {
 };
 
 exports.findDraftedPostsByUserId = (req, res) => {
-  const id = req.body.userId;
+  if ((req.session==undefined || req.session.user==undefined)) {
+    res.status(400).send({
+      message: "Please login first"
+    });
+    return;
+  }
+  const id = req.session.user.id;
   posts.findAll({ where: { userId: id, published: false } })
     .then((post) => {
       if (post) {
@@ -90,6 +114,12 @@ exports.findDraftedPostsByUserId = (req, res) => {
 
 exports.DeletePost = (req, res) => {
   const pid = req.body.postId;
+  if ((req.session==undefined || req.session.user==undefined)) {
+    res.status(400).send({
+      message: "Please login first"
+    });
+    return;
+  }
   posts.destroy({ where: { id: pid } })
     .then((post) => {
       if (post == 1) {
@@ -110,6 +140,12 @@ exports.DeletePost = (req, res) => {
 
 exports.UpdatePost = (req, res) => {
   const pid = req.body.postId;
+  if ((req.session==undefined || req.session.user==undefined)) {
+    res.status(400).send({
+      message: "Please login first"
+    });
+    return;
+  }
   posts.update(req.body, { where: { id: pid } })
     .then(num => {
       if (num == 1) {
